backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
asyncpg==0.27.0
python-dotenv==1.0.0
jinja2==3.1.2
